requirements.txt
--index-url https://download.pytorch.org/whl/cu126
torch
torchvision
torchaudio

transformers
datasets
peft
bitsandbytes
accelerate
safetensors
sentencepiece

pillow
requests
tqdm
huggingface_hub

fastapi
uvicorn[standard]
pydantic
